feat(1830): build cache to disk
