+++ lib/buildFactory.js
@@ -3,85 +3,9 @@
 const BaseFactory = require('./baseFactory');
 const Build = require('./build');
 const imageParser = require('docker-parse-image');
-const buildClusterAnnotation = 'screwdriver.cd/buildCluster';
+const getBuildClusterName = require('./helper').getBuildClusterName;
 const hoek = require('hoek');
 let instance;
-
-/**
- * Pick a random cluster from the array based on their weightage
- * @method getRandomCluster
- * @param  {Array}          clusters    An array of build clusters
- * @return {String}                     Name of the picked cluster
- */
-function getRandomCluster(clusters) {
-    // Add up all the weightage, and generate a random number between (0, totalWeight - 1)
-    const totalWeight = clusters.reduce((prev, cur) => prev + cur.weightage, 0);
-    const number = Math.floor(Math.random() * totalWeight);
-    let sum = 0;
-
-    for (let i = 0; i < clusters.length; i += 1) {
-        sum += clusters[i].weightage;
-
-        if (number < sum) return clusters[i].name;
-    }
-
-    return clusters[0].name;
-}
-
-/**
- * Get build cluster name based on annotations
- * @method getBuildClusterName
- * @param  {Object}         config
- * @param  {Object}         config.annotations              Annotations for the job
- * @param  {PipelineModel}  config.pipeline                 Pipeline model
- * @return {String}        Build cluster name
- */
-async function getBuildClusterName({ annotations, pipeline }) {
-    const buildClusterName = annotations[buildClusterAnnotation] || '';
-    // Lazy load factory dependency to prevent circular dependency issues
-    // https://nodejs.org/api/modules.html#modules_cycles
-    // eslint-disable-next-line global-require
-    const BuildClusterFactory = require('./buildClusterFactory');
-    const buildClusterFactory = BuildClusterFactory.getInstance();
-
-    if (!buildClusterName) {
-        return buildClusterFactory.list({
-            params: {
-                managedByScrewdriver: true,
-                isActive: true
-            }
-        }).then((buildClusters) => {
-            if (buildClusters.length === 0) {
-                return '';
-            }
-
-            return getRandomCluster(buildClusters);
-        });
-    }
-
-    return buildClusterFactory.get({
-        name: buildClusterName
-    }).then((buildCluster) => {
-        if (!buildCluster) {
-            // eslint-disable-next-line max-len
-            throw new Error(`Cluster specified in screwdriver.cd/buildCluster ${buildClusterName} does not exist.`);
-        }
-
-        // Check if this pipeline's org is authorized to use the build cluster
-        // pipeline name example: screwdriver-cd/ui
-        const regex = /^([^/]+)\/.*/;
-        const matched = pipeline.name.match(regex);
-        const org = matched ? matched[1] : '';
-
-        if (buildCluster.scmContext !== pipeline.scmContext
-            || (buildCluster.managedByScrewdriver === false
-                && !buildCluster.scmOrganizations.includes(org))) {
-            throw new Error('This pipeline is not authorized to use this build cluster.');
-        }
-
-        return buildCluster.name;
-    });
-}
 
 /**
  * Gathers a sha for the build
