+++ lib/pipeline.js
@@ -11,6 +11,7 @@ const winston = require('winston');
 const dayjs = require('dayjs');
 const _ = require('lodash');
 const { getAllRecords } = require('./helper.js');
+const getBuildClusterName = require('./helper').getBuildClusterName;
 const { PR_JOB_NAME, EXTERNAL_TRIGGER } = require('screwdriver-data-schema').config.regex;
 const REGEX_CAPTURING_GROUP = {
     pr: 1, // PR-1
@@ -1187,9 +1188,24 @@ class PipelineModel extends BaseModel {
                 scmContext: this.scmContext,
                 token
             }))
-            .then((scmRepo) => {
+            .then(async (scmRepo) => {
                 this.scmRepo = scmRepo;
                 this.name = scmRepo.name;
+
+                let buildClusterName;
+                const annotations = this.annotations;
+                const config = this;
+
+                if (String(this.multiBuildClusterEnabled) === 'true') {
+                    buildClusterName = await getBuildClusterName({
+                        annotations,
+                        config
+                    });
+                }
+
+                if (buildClusterName) {
+                    this.buildClusterName = buildClusterName;
+                }
 
                 return super.update();
             });
