+++ lib/pipelineFactory.js
@@ -2,6 +2,7 @@
 
 const BaseFactory = require('./baseFactory');
 const Pipeline = require('./pipeline');
+const getBuildClusterName = require('./helper').getBuildClusterName;
 let instance;
 
 class PipelineFactory extends BaseFactory {
@@ -11,10 +12,12 @@ class PipelineFactory extends BaseFactory {
      * @param  {Object}    config
      * @param  {Object}    config.datastore         Object that will perform operations on the datastore
      * @param  {Boolean}   config.externalJoin      Flag to allow external join
+     * @param  {Boolean}   config.multiBuildClusterEnabled  Enable multiple build cluster feature or not
      */
     constructor(config) {
         super('pipeline', config);
         this.externalJoin = config.externalJoin || false;
+        this.multiBuildClusterEnabled = config.multiBuildClusterEnabled;
     }
 
     /**
@@ -63,9 +66,23 @@ class PipelineFactory extends BaseFactory {
                 scmContext: config.scmContext,
                 token
             }))
-            .then((scmRepo) => {
+            .then(async (scmRepo) => {
                 modelConfig.scmRepo = scmRepo;
                 modelConfig.name = scmRepo.name;
+
+                let buildClusterName;
+                const annotations = config.annotations;
+
+                if (String(this.multiBuildClusterEnabled) === 'true') {
+                    buildClusterName = await getBuildClusterName({
+                        annotations,
+                        config
+                    });
+                }
+
+                if (buildClusterName) {
+                    modelConfig.buildClusterName = buildClusterName;
+                }
 
                 return super.create(modelConfig);
             });
